==> notebooks/4_qa_analyze/02_derive_neurosynth_terms.ipynb <==
%load_ext autoreload
%autoreload 2
import sys
from ridge_utils.DataSequence import DataSequence
import pandas as pd
import os
import matplotlib.pyplot as plt
import cortex
import seaborn as sns
from os.path import join
from collections import defaultdict
import numpy as np
import joblib
import dvu
import json
import sys
sys.path.append('../notebooks')
from tqdm import tqdm
from neuro import analyze_helper, viz
from neuro.features.qa_questions import get_questions, get_merged_questions_v3_boostexamples

# https://github.com/neurosynth/neurosynth-data/blob/master/data-neurosynth_version-7_vocab-terms_vocabulary.txt

df = pd.read_csv('data-neurosynth_version-7_vocab-terms_vocabulary.txt')

vals = df['001'].str.strip()
assert len(vals) == len(vals.unique())
print(len(vals))
# vals remove any terms that contain numbers
vals = vals[~vals.str.contains(r'\d')]
print(len(vals))
vals = vals[vals.str.len() > 2]
print(len(vals))

# remove val if val + 's' in vals
vals = vals[~vals.apply(lambda x: (x + 's') in vals.values)]
print(len(vals))
vals = vals[~vals.apply(lambda x: (x + 'ly') in vals.values)]
print(len(vals))

vals.to_csv('terms_full.csv')

prompt = '''You are given a comma-separated list of words / phrases. Your task is to classify whether each element of the list is a semantic text property or not. Some of the words instead describe properties of brain regions or diseases. Answer yes or no and return your answer in json.


### Example input
Input: abstract, acc, add, addiction, alcohol, anger, arithmetic, baseline, experiments, family, foot

### Example output
```
{
  "abstract": "yes",
  "acc": "no",
  "addiction": "yes",
  "alcohol": "no",
  "anger": "yes",
  "arithmetic": "yes",
  "baseline": "no",
  "date": "yes",
  "debate": "yes",
  "dti": "no",
  "exhibit": "no",
  "experiments": "no",
  "family": "yes",
  "foot": "yes"
}
```

### Input
'''

', '.join(vals.values)

filt1 = json.load(open('terms_filt1.json', 'r'))

vals_filt1 = sorted({k for k in filt1 if filt1[k] == 'yes'})

pd.DataFrame(vals_filt1).to_csv('terms_filt1.csv')

len(vals), len(vals_filt1)

', '.join(vals_filt1)
